requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
boto3==1.35.91
botocore==1.35.91
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.6.7
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.1.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.12.0
jedi==0.19.2
jmespath==1.0.1
jupyter-client==7.3.4
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
nest_asyncio==1.6.0
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==24.2
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
s3transfer==0.10.4
selenium==4.27.1
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack_data==0.6.3
tornado==6.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
